backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
supabase
httpx
requests
pytest
pytest-asyncio
supabase
python-dotenv
pydantic
websockets
